Rename common module
